+++ frontend/src/pages/AdminPage/ui/AdminArticles/AdminArticles.tsx
@@ -1,0 +1,7 @@
+export const AdminArticles = () =>{
+    return(
+        <div>
+            Здесь будут статьи
+        </div>
+    )
+}
